abstracted functions for updating form values, including initialization and batch updates. Also added a new higher-order component, ArgInput.
